refactor(failover): consolidate QuickNode failover config into one module
Move the QuickNode map, getFailoverUrlsForInfuraNetwork, and
getIsQuicknodeEndpointUrl into network-failover.ts so all the failover config
lives in one place, and switch FEATURED_RPCS to the chain keyed
getFailoverUrlsForChainId(CHAIN_IDS.X) so you can tell at a glance which chain
each entry refers to.

To break the network.ts <-> network-failover.ts cycle this would create,
CHAIN_IDS moves into its own leaf module (chain-ids.ts) that both import;
network.ts re-exports CHAIN_IDS plus the moved failover symbols so existing
importers, including the frozen migrations, are untouched.

No behavior change. Circular deps check passes.
