+++ shared/constants/chain-ids.ts
@@ -1,0 +1,126 @@
+/**
+ * An object containing all of the chain ids for networks both built in and
+ * those that we have added custom code to support our feature set.
+ *
+ * This lives in its own leaf module (no imports) so that other constant modules
+ * such as `network-failover.ts` can reference chain ids without importing the
+ * large `network.ts`, which would create a circular import. `network.ts`
+ * re-exports `CHAIN_IDS`, so existing importers are unaffected.
+ */
+export const CHAIN_IDS = {
+  MAINNET: '0x1',
+  GOERLI: '0x5',
+  LOCALHOST: '0x539',
+  BSC: '0x38',
+  BSC_TESTNET: '0x61',
+  OPTIMISM: '0xa',
+  OPTIMISM_TESTNET: '0xaa37dc',
+  OPTIMISM_GOERLI: '0x1a4',
+  BASE: '0x2105',
+  BASE_TESTNET: '0x14a33',
+  OPBNB: '0xcc',
+  OPBNB_TESTNET: '0x15eb',
+  POLYGON: '0x89',
+  POLYGON_TESTNET: '0x13881',
+  AVALANCHE: '0xa86a',
+  AVALANCHE_TESTNET: '0xa869',
+  FANTOM: '0xfa',
+  FANTOM_TESTNET: '0xfa2',
+  CELO: '0xa4ec',
+  ARBITRUM: '0xa4b1',
+  HARMONY: '0x63564c40',
+  PALM: '0x2a15c308d',
+  SEPOLIA: '0xaa36a7',
+  HOLESKY: '0x4268',
+  LINEA_GOERLI: '0xe704',
+  LINEA_SEPOLIA: '0xe705',
+  AMOY: '0x13882',
+  BASE_SEPOLIA: '0x14a34',
+  BLAST_SEPOLIA: '0xa0c71fd',
+  OPTIMISM_SEPOLIA: '0xaa37dc',
+  PALM_TESTNET: '0x2a15c3083',
+  CELO_TESTNET: '0xaef3',
+  ZK_SYNC_ERA_TESTNET: '0x12c',
+  MANTA_SEPOLIA: '0x138b',
+  UNICHAIN: '0x82',
+  UNICHAIN_SEPOLIA: '0x515',
+  LINEA_MAINNET: '0xe708',
+  AURORA: '0x4e454152',
+  MOONBEAM: '0x504',
+  MOONBEAM_TESTNET: '0x507',
+  MOONRIVER: '0x505',
+  CRONOS: '0x19',
+  GNOSIS: '0x64',
+  ZKSYNC_ERA: '0x144',
+  TEST_ETH: '0x539',
+  ARBITRUM_GOERLI: '0x66eed',
+  BLAST: '0x13e31',
+  FILECOIN: '0x13a',
+  POLYGON_ZKEVM: '0x44d',
+  SCROLL: '0x82750',
+  SCROLL_SEPOLIA: '0x8274f',
+  WETHIO: '0x4e',
+  CHZ: '0x15b38',
+  NUMBERS: '0x290b',
+  SEI: '0x531',
+  APECHAIN_TESTNET: '0x8157',
+  APECHAIN_MAINNET: '0x8173',
+  BERACHAIN: '0x138de',
+  METACHAIN_ONE: '0x1b6e6',
+  ARBITRUM_SEPOLIA: '0x66eee',
+  NEAR: '0x18d',
+  NEAR_TESTNET: '0x18e',
+  B3: '0x208d',
+  B3_TESTNET: '0x7c9',
+  GRAVITY_ALPHA_MAINNET: '0x659',
+  GRAVITY_ALPHA_TESTNET_SEPOLIA: '0x34c1',
+  LISK: '0x46f',
+  LISK_SEPOLIA: '0x106a',
+  INK_SEPOLIA: '0xba5eD',
+  INK: '0xdef1',
+  GSYS_MAINNET: '0x407b',
+  GSYS_TESTNET: '0xa5c8',
+  MODE_SEPOLIA: '0x397',
+  MODE: '0x868b',
+  MEGAETH_TESTNET: '0x18c6',
+  MEGAETH_TESTNET_V2: '0x18c7',
+  MEGAETH_MAINNET: '0x10e6',
+  XRPLEVM_TESTNET: '0x161c28',
+  LENS: '0xe8',
+  PLUME: '0x18232',
+  MATCHAIN: '0x2ba',
+  FLOW: '0x2eb',
+  KATANA: '0xb67d2',
+  MONAD_TESTNET: '0x279f',
+  SOPHON: '0xc3b8',
+  SOPHON_TESTNET: '0x1fa72e78',
+  EDUCHAIN: '0xa3c3',
+  ABSTRACT: '0xab5',
+  NOMINA: '0xa6',
+  XDC: '0x32',
+  XRPLEVM: '0x15f900',
+  FRAX: '0xfc',
+  ACALA: '0x313',
+  ACALA_TESTNET: '0x253',
+  KARURA: '0x2ae',
+  HEMI: '0xa867',
+  PLASMA: '0x2611',
+  LUKSO: '0x2a',
+  INJECTIVE: '0x6f0',
+  MONAD: '0x8f',
+  HYPE: '0x3e7',
+  X_LAYER: '0xc4',
+  ETHERLINK: '0xa729',
+  MSU: '0x10b3e',
+  BOB: '0xed88',
+  ROOTSTOCK: '0x1e',
+  ROOTSTOCK_TESTNET: '0x1f',
+  TEMPO_TESTNET: '0xa5bf',
+  TEMPO_MAINNET: '0x1079',
+  STABLE_MAINNET: '0x3dc',
+  MANTLE: '0x1388',
+  KONET: '0x4341',
+  ARC: '0x13b2',
+  ROBINHOOD_CHAIN: '0x1237',
+  STABLE: '0x3dc',
+} as const;
